metric, closes #146: added retry logic to writing metrics to cw;
If writing to cloud watch failes, we log an error and thus raise an
alarm. However, neither is this useful nor necessary as we can retry
writing the metric just fine. So this commit introducing a retry logic,
but needs to refactor our package layout a little bit as mon and cloud
can not depend on each other.
